requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
ccxt==4.4.69
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
idna==3.10
joblib==1.4.2
multidict==6.2.0
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
propcache==0.3.0
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
